Add sidebar position
Sidebar position places dmenu on the left side of the screen, filling it
vertically. minwidth/-mw is used to determine the minimum width as with
center position. Sidebar position will automatically set lines to the
appropriate value so that as many items as possible are listed.

Also fix man page.